// src/models/content-manangement/category/category.response.ts
export interface AllCategories {
  content: CategoryModel[];
  pageNo: number;
  pageSize: number;
  totalElements: number;
  totalPages: number;
  last: boolean;
  first: boolean;
  hasNext: boolean;
  hasPrevious: boolean;
}

export interface CategoryModel {
  id: string;
  createdAt: string;
  updatedAt: string;
  createdBy: string;
  updatedBy: string;
  businessId: string;
  businessName: string;
  name: string;
  imageUrl: string;
  status: string;
  totalProducts: number;
}